move some code around
